requirements.txt
nicegui==1.2.16
pydantic==1.10.8
python-dotenv==1.0.0